requirements.txt
evtx~=0.8.2
PyYAML~=6.0.1